requirements.txt
PyPDF2
python-docx
google-generativeai
pillow
moviepy
numpy
opencv-python
streamlit